Support for paths in GitLabSecurityDashboard (#266)
* Additional fix for the Absolute_Path_Traversal vulnerability: support for paths in GitLabSecurityDashboard instead of just filenames.
* Extracted duplicate code in issue trackers into a base class
