+++ src/main/java/com/checkmarx/flow/service/FilenameFormatter.java
@@ -13,5 +13,12 @@ public interface FilenameFormatter {
      */
     String formatPath(ScanRequest request, String filenameTemplate, String baseDir);
 
+    /**
+     * Formats filenameTemplate by replacing placeholders in the template with values from request.
+     * Additionally, current timestamp may be specified as a placeholder.
+     * @param request used as a source of placeholder values
+     * @param filenameTemplate filename that may optionally contain placeholders
+     * @return the formatted filename
+     */
     String formatFilenameTemplate(ScanRequest request, String filenameTemplate);
 }
